+++ test/unit/specs/components/image-info.spec.js
@@ -1,11 +1,9 @@
-import ImageInfo from '@/components/ImageInfo';
+import ImageInfo from '@/components/ImageDetails/ImageInfo';
 import render from '../../test-utils/render';
 
 describe('Image Info', () => {
   let props = null;
   let options = {};
-  let storeState = null;
-  let commitMock = null;
 
   beforeEach(() => {
     props = {
@@ -27,22 +25,8 @@ describe('Image Info', () => {
       imageWidth: 500,
     };
 
-    commitMock = jest.fn();
-
-    storeState = {
-      $store: {
-        commit: commitMock,
-        state: {
-          isReportFormVisible: false,
-        },
-      },
-    };
-
     options = {
       propsData: props,
-      mocks: {
-        ...storeState,
-      },
     };
   });
 
@@ -67,26 +51,5 @@ describe('Image Info', () => {
     const wrapper = render(ImageInfo, options);
     expect(wrapper.html()).toContain(`${props.imageWidth}`);
     expect(wrapper.html()).toContain(`${props.imageHeight} pixels`);
-  });
-
-  it('should toggle visibility of report form on report button click', () => {
-    const wrapper = render(ImageInfo, options);
-    const button = wrapper.find('.report');
-    button.trigger('click');
-
-    expect(commitMock).toHaveBeenCalledWith('TOGGLE_REPORT_FORM_VISIBILITY');
-  });
-
-  it(' report form should be invisible by default', () => {
-    const wrapper = render(ImageInfo, options);
-
-    expect(wrapper.find({ name: 'content-report-form' }).vm).not.toBeDefined();
-  });
-
-  it(' report form should be visible when isReportFormVisible is true', () => {
-    storeState.$store.state.isReportFormVisible = true;
-    const wrapper = render(ImageInfo, options);
-
-    expect(wrapper.find({ name: 'content-report-form' }).vm).toBeDefined();
   });
 });
