+++ test/unit/specs/components/photo-details.spec.js
@@ -1,9 +1,11 @@
-import PhotoDetails from '@/components/PhotoDetails';
+import PhotoDetails from '@/components/ImageDetails/PhotoDetails';
 import render from '../../test-utils/render';
 
 describe('PhotoDetails', () => {
   let options = null;
   let props = null;
+  let storeState = null;
+  let commitMock = null;
 
   beforeEach(() => {
     props = {
@@ -22,8 +24,22 @@ describe('PhotoDetails', () => {
       socialSharingEnabled: true,
     };
 
+    commitMock = jest.fn();
+
+    storeState = {
+      $store: {
+        commit: commitMock,
+        state: {
+          isReportFormVisible: false,
+        },
+      },
+    };
+
     options = {
       propsData: props,
+      mocks: {
+        ...storeState,
+      },
     };
   });
 
@@ -64,22 +80,14 @@ describe('PhotoDetails', () => {
   });
 
   it('renders link back to search results if enabled', () => {
-    const wrapper = render(PhotoDetails, {
-      propsData: {
-        ...props,
-        shouldShowBreadcrumb: true,
-      },
-    });
+    options.propsData.shouldShowBreadcrumb = true;
+    const wrapper = render(PhotoDetails, options);
     expect(wrapper.find('.photo_breadcrumb').element).toBeDefined();
   });
 
   it('doesnt render link back to search results if disabled', () => {
-    const wrapper = render(PhotoDetails, {
-      propsData: {
-        ...props,
-        shouldShowBreadcrumb: false,
-      },
-    });
+    options.propsData.shouldShowBreadcrumb = false;
+    const wrapper = render(PhotoDetails, options);
     expect(wrapper.find('.photo_breadcrumb').element).toBeUndefined();
   });
 
@@ -103,11 +111,33 @@ describe('PhotoDetails', () => {
       mocks: {
         $router: routerMock,
         $route: routeMock,
+        ...storeState,
       },
     };
     const wrapper = render(PhotoDetails, opts);
     const link = wrapper.find('.photo_breadcrumb');
     link.trigger('click');
     expect(routerMock.push).toHaveBeenCalledWith({ name: 'browse-page', query: opts.propsData.query, params: { location: routeMock.params.location } });
+  });
+
+  it('should toggle visibility of report form on report button click', () => {
+    const wrapper = render(PhotoDetails, options);
+    const button = wrapper.find('.report');
+    button.trigger('click');
+
+    expect(commitMock).toHaveBeenCalledWith('TOGGLE_REPORT_FORM_VISIBILITY');
+  });
+
+  it(' report form should be invisible by default', () => {
+    const wrapper = render(PhotoDetails, options);
+
+    expect(wrapper.find({ name: 'content-report-form' }).vm).not.toBeDefined();
+  });
+
+  it(' report form should be visible when isReportFormVisible is true', () => {
+    storeState.$store.state.isReportFormVisible = true;
+    const wrapper = render(PhotoDetails, options);
+
+    expect(wrapper.find({ name: 'content-report-form' }).vm).toBeDefined();
   });
 });
